Fix help command
Help command was showing the previous AB3
user guide. Due to the changes we made to the
AB3, we had to update the user guide and therefore
update to the new link.

Now with the new link, users who need help in the
address book can look at our updated user guide.