Fix invalid message for edit.
Last time, we could edit role. However, we changed
the behaviour so that we cannot edit role. Therefore,
we must remove any residue from that time. edit 1 r/
was showing that the field cannot be empty instead of an
invalid command format despite it being an invalid command
format.

Now when users try to edit role using r/, the system will show that
its an invalid command format.